# Optimize FindMiniHeaders by using gob encoding for event logs