@albrow 2020-06-16
@jalextowle should we include these new functions in our conversion tests? If so, what is the best way to do that? This is the first time in the db package where we actually do some hand-written conversion to `JSValue` as opposed to just using `jsutil.InefficientlyConvertToJS`.