@albrow 2020-06-17
@jalextowle added a new test baed on what we talked about. Can you take a look at https://github.com/0xProject/0x-mesh/pull/840/commits/e8afa0d4e2f9e7b460f8ea37aba61f50bacda3df and see if that is sufficient?