@albrow 2020-06-17
@jalextowle

> Should this be “gobMiniHeader”, or is “got” the prefix that you intended?

It was intended to be `gotMiniHeader` but I renamed it to make it more clear.